backend/requirements.txt
asgiref==3.9.1
click==8.3.1
colorama==0.4.6
dj-database-url==3.1.2
Django==4.2.28
django-cors-headers==4.7.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==25.1.0
h11==0.16.0
packaging==26.0
psycopg[binary]==3.2.12
PyJWT==2.10.1
python-dotenv==1.2.1
sqlparse==0.5.3
tzdata==2025.2
uvicorn==0.41.0
whitenoise==6.12.0
